Extract topic store health module
